[ft #156063977] Add pagination for categories and recipes
	public/index.html - Add Bootstrap 3 cdn for react-bootstrap
	src/App/App.jsx - Remove css class distorting display after adding Boostrap 3
	src/Dashboard/Dashboard.jsx - Add react bootstrap pagination functionality
	src/RecipePage/RecipePage.jsx - Rename category id variable and recipe ingredients
	src/_actions/category.actions.js - Pass page number variable to the view category action creator
	src/_actions/recipe.actions.js - Pass page number variable to the recipe action creator
	src/_components/RecipeForm.js - Change recipe description variable to ingredient
	src/_reducers/category.reducer.js - Edit validation error messages
	src/_reducers/category.reducer.js - Edit how the payload passed to state
	src/_reducers/recipe.reducer.js - Edit how the payload passed to state
	src/_services/category.services.js - Edit the axios calls to cater for page numbers
	src/_services/recipe.services.js - Edit the axios calls to cater for page numbers
	src/styles/css/style.css - Added css for the nav bar and adjustments to cater for distortion after adding bootstrap 3
	src/_reducers/authentication.reducer.js -Add authentication reducer

+++ src/Dashboard/Dashboard.jsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import Collapsible from 'react-collapsible';
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
+import { Pagination } from 'react-bootstrap';
 
 import Layout from '../Layout/Layout';
 import { CategoryPage } from '../CategoryPage/CategoryPage';
@@ -20,6 +23,8 @@ class Dashboard extends React.Component {
       recipes: [],
       categoryOpen: false,
       recipeOpen: false,
+      currentCategoryPage: 1,
+      currentRecipePage: 1,
     };
   }
   getCategory = (category) => {
@@ -37,7 +42,22 @@ class Dashboard extends React.Component {
   };
 
   deleteACategory = (category) => {
-    this.props.deleteCategory(category);
+    confirmAlert({
+      title: 'Confirm to submit',
+      message: 'Are you sure to do this.',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            this.props.deleteCategory(category);
+          },
+        },
+        {
+          label: 'No',
+          onClick: () => {},
+        },
+      ],
+    });
   };
 
   viewACategoriesRecipes = (categoryId) => {
@@ -45,11 +65,137 @@ class Dashboard extends React.Component {
   };
 
   deleteARecipe = (recipe) => {
-    this.props.deleteRecipe(recipe);
+    confirmAlert({
+      message: 'Are you sure you want to delete this recipe',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            this.props.deleteRecipe(recipe);
+          },
+        },
+        {
+          label: 'No',
+          onClick: () => {},
+        },
+      ],
+    });
+  };
+  selectedCategoryPage = (pageNum) => {
+    console.log('this is the category page --->>>>>', pageNum);
+    this.props.viewCategory(pageNum);
+    this.setState({ currentCategoryPage: pageNum });
+  };
+
+  selectedRecipePage = (pageNum) => {
+    console.log('this is the recipe page --->>>>>', pageNum);
+    this.props.viewRecipes(pageNum);
+    this.setState({ currentRecipePage: pageNum });
+  };
+  categoryPagination = (categoryPages) => {
+    const theItems = [];
+    if (categoryPages) {
+      theItems.push( 
+        <Pagination.First
+          key={0}
+          disabled={this.state.currentCategoryPage === 1}
+          onClick={this.selectedCategoryPage.bind(this, 1)}
+        />,
+        <Pagination.Prev
+          key={1}
+          disabled={this.state.currentCategoryPage === 1}
+          onClick={this.selectedCategoryPage.bind(
+            this,
+            this.state.currentCategoryPage - 1 > 1 ? this.state.currentCategoryPage - 1 : 1,
+          )}
+        />,
+      );
+
+      for (let i = 1; i <= categoryPages; i++) {
+        theItems.push(
+          <Pagination.Item
+            key={i + 1}
+            active={i === this.state.currentCategoryPage}
+            onClick={this.selectedCategoryPage.bind(this, i)}
+          >
+            {i}
+          </Pagination.Item>,
+        );
+      }
+      theItems.push(
+        <Pagination.Next
+          key={theItems.length + 1}
+          disabled={this.state.currentCategoryPage === categoryPages}
+          onClick={this.selectedCategoryPage.bind(
+            this,
+            this.state.currentCategoryPage + 1 <= categoryPages
+              ? this.state.currentCategoryPage + 1
+              : categoryPages,
+          )}
+        />,
+        <Pagination.Last
+          key={theItems.length + 2}
+          disabled={this.state.currentCategoryPage === categoryPages}
+          onClick={this.selectedCategoryPage.bind(this, categoryPages)}
+        />,
+      );
+    }
+    return theItems;
+  };
+
+  recipePagination = (recipePages) => {
+    const theItems = [];
+    if (recipePages) {
+      theItems.push(
+        <Pagination.First
+          key={0}
+          disabled={this.state.currentRecipePage === 1}
+          onClick={this.selectedRecipePage.bind(this, 1)}
+        />,
+        <Pagination.Prev
+          key={1}
+          disabled={this.state.currentRecipePage === 1}
+          onClick={this.selectedRecipePage.bind(
+            this,
+            this.state.currentRecipePage - 1 > 1 ? this.state.currentRecipePage - 1 : 1,
+          )}
+        />,
+      );
+
+      for (let i = 1; i <= recipePages; i++) {
+        theItems.push(
+          <Pagination.Item
+            key={i + 1}
+            active={i === this.state.currentRecipePage}
+            onClick={this.selectedRecipePage.bind(this, i)}
+          >
+            {i}
+          </Pagination.Item>,
+        );
+      }
+      theItems.push(
+        <Pagination.Next
+          key={theItems.length + 1}
+          disabled={this.state.currentRecipePage === recipePages}
+          onClick={this.selectedRecipePage.bind(
+            this,
+            this.state.currentRecipePage + 1 <= recipePages
+              ? this.state.currentRecipePage + 1
+              : recipePages,
+          )}
+        />,
+        <Pagination.Last
+          key={theItems.length + 2}
+          disabled={this.state.currentRecipePage === recipePages}
+          onClick={this.selectedRecipePage.bind(this, recipePages)}
+        />,
+      );
+    }
+    return theItems;
   };
 
   render() {
-    const { categories, recipes } = this.props;
+    const { categories, recipes, categoryPages, recipePages } = this.props;
     const { categoryOpen, recipeOpen } = this.state;
     console.log('_-_-_-_-_-_-_-_-_-', this.state.categoryOpen);
     console.log('*-*-*-*-*-*-*-*-*-', this.state.recipeOpen);
@@ -58,7 +204,7 @@ class Dashboard extends React.Component {
         <div>
           <br />
           <br />
-          <p>You're logged into Yummy Recipes</p>
+          {/* <p>You're logged into Yummy Recipes</p> */}
           <h3>Welcome</h3>
         </div>
         <div className="row">
@@ -114,7 +260,7 @@ class Dashboard extends React.Component {
                                   onClick={() => this.deleteACategory(category)}
                                   className="btn-lg btn-danger btn-sm"
                                 >
-                                  <i className="fa fa-trash-o" aria-hidden="true" />
+                                  <i className="fa fa-trash-o mr-2" aria-hidden="true" />
                                 </button>
                               </div>
                             </td>
@@ -128,7 +274,13 @@ class Dashboard extends React.Component {
             ) : (
               <div> No Categories </div>
             )}
+            <div className="container-fluid">
+              <div className="text-centered">
+                <Pagination bsSize="medium"> {this.categoryPagination(categoryPages)} </Pagination>
+              </div>
+            </div>
           </div>
+
           {/* Recipes */}
           <div className="col-sm-8">
             <div>
@@ -183,6 +335,13 @@ class Dashboard extends React.Component {
                       ) : (
                         <div> No recipes </div>
                       )}
+                      <div className="container-fluid">
+                        <div className="text-centered">
+                          <Pagination bsSize="medium">
+                            {this.recipePagination(recipePages)}
+                          </Pagination>
+                        </div>
+                      </div>
                     </div>
                   </div>
                   <div>{!!recipeOpen && <EditRecipe />}</div>
@@ -197,11 +356,26 @@ class Dashboard extends React.Component {
 }
 
 function mapStateToProps(state) {
-  console.log('@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@>>>>>', state.viewCategoryReducer.categories);
+  const { categories, categoryPages, categoryPage } = state.viewCategoryReducer;
+  const { recipes, recipePages, recipePage } = state.viewRecipeReducer;
+  console.log(
+    '@@@@@@@@@@@@################++++++++++++++++>>>>>',
+    categories,
+    categoryPages,
+    categoryPage,
+    recipes,
+    recipePages,
+    recipePage,
+  );
+  console.log('@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@>>>>>', state.categories);
   console.log('@@@@@@@@@@@@@@@@@@@recipes@@@@@@@@@@@@@@@@@>>>>>', state.viewRecipeReducer.recipes);
   return {
-    categories: state.viewCategoryReducer.categories,
-    recipes: state.viewRecipeReducer.recipes,
+    categories: categories,
+    recipes: recipes,
+    categoryPages: categoryPages,
+    recipePages: recipePages,
+    currentCategoryPage: categoryPage,
+    currentRecipePage: recipePage,
   };
 }
